backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
aiosqlite
chromadb
httpx
pytest
pytest-asyncio
